add pixel-interpolator test cases
The pixel interpolator used by e.g. Perspective Correction depends on
a setting in darktablerc.  To have regression tests for all four
existing interpolators, this commit adds code to run.sh which permits
adding a --conf flag to the invocation of darktable-cli, as well as
three new tests which are the same as 0018 but with differing
interpolators rather than the default bicubic.

Co-authored-by: Pascal Obry <pascal@obry.net>